Refactor(twins): split network.go classes into files and update docs.
